pyproject.toml
[tool.poetry]
name = "jet-profile-cleaner"
version = "0.1.0"
description = ""
authors = ["Doctor <thirvondukr@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = {git = "https://github.com/Rapptz/discord.py.git"}
python-dotenv = "^0.20.0"
pydantic = "^1.9.2"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
